requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
email-validator==2.2.0
google-cloud-firestore==2.21.0
